#106 test : 지역 목록 삭제시 대표지역 설정 테스트, 구조 리팩토링
